feat: Discord Bot機能をmainに統合．画像読み込みから文字列抽出までの一連を実装．
